Restored proper tense in limit_conn description.
The limit_conn module only limits connections with active requests, that
is, if a connection has a request "being processed" by the server.  During
a text review in 95c3c3bbf1ce it was changed to "has a request processed",
which is incorrect (it is true after processing the first request).

Reported by Abilio Marques,
http://mailman.nginx.org/pipermail/nginx/2017-December/055322.html.